added file to forked repo
I tried to recreate the step function used in the paper (Deep Residual Learning for Nonlinear Regression. Entropy 2020, 22, 193.) The data are created using random generator & python function, so no need to load a csv file. I ran it for about 100k or 1mil rows instead of 10 mil to save time because the result won't be too different (for 10 mil it takes about 20 minutes).